Move RecordReceiverContext into its own file.
