+++ components/core/src/reducer/reducer_server.cpp
@@ -11,92 +11,13 @@
 
 #include "../clp/spdlog_with_specializations.hpp"
 #include "CommandLineArguments.hpp"
-#include "DeserializedRecordGroup.hpp"
+#include "RecordReceiverContext.hpp"
 #include "ServerContext.hpp"
 #include "types.hpp"
 
 using boost::asio::ip::tcp;
 
 namespace reducer { namespace {
-class RecordReceiverContext {
-public:
-    static constexpr size_t cMinBufSize = 1024;
-
-    explicit RecordReceiverContext(std::shared_ptr<ServerContext> const& ctx)
-            : m_server_ctx(ctx),
-              m_socket(ctx->get_io_context()),
-              m_buf(cMinBufSize) {}
-
-    ~RecordReceiverContext() { m_socket.close(); }
-
-    // Disallow copy and move
-    RecordReceiverContext(RecordReceiverContext const&) = delete;
-    RecordReceiverContext(RecordReceiverContext const&&) = delete;
-    RecordReceiverContext const& operator=(RecordReceiverContext const&) = delete;
-    RecordReceiverContext const& operator=(RecordReceiverContext const&&) = delete;
-
-    static std::shared_ptr<RecordReceiverContext> new_receiver(
-            std::shared_ptr<ServerContext> const& ctx
-    ) {
-        auto receiver = std::make_shared<RecordReceiverContext>(ctx);
-
-        // Clear the v6_only flag to allow ipv4 and ipv6 connections, but only on Linux. For full
-        // portability, we need separate v4 and v6 acceptors.
-        // boost::asio::ip::v6_only option(false);
-        // receiver->m_socket.set_option(option);
-
-        return receiver;
-    }
-
-    /**
-     * Reads a connection initiation packet.
-     * @return false if there are an unexpected number of bytes in the buffer or the sender's job ID
-     * doesn't match the one currently being processed.
-     * @return true otherwise.
-     */
-    bool read_connection_init_packet();
-
-    /**
-     * Sends a connection accept packet.
-     * @return Whether the acceptance was sent successfully.
-     */
-    bool send_connection_accept_packet();
-
-    /**
-     * Reads a packet containing record groups.
-     * @return Whether the read was successful.
-     */
-    bool read_record_groups_packet();
-
-    /**
-     * NOTE: This method should only be called from handlers for boost::asio read tasks on the
-     * receiver's socket.
-     * @param num_bytes
-     */
-    void increment_buf_num_bytes_occupied(size_t num_bytes) {
-        m_buf_num_bytes_occupied += num_bytes;
-    }
-
-    std::shared_ptr<ServerContext>& get_server_ctx() { return m_server_ctx; }
-
-    tcp::socket& get_socket() { return m_socket; }
-
-    /**
-     * @return A pointer to the next writable byte in the buffer.
-     */
-    char* get_buf_write_head() { return &m_buf[m_buf_num_bytes_occupied]; }
-
-    size_t get_buf_num_bytes_avail() { return m_buf.size() - m_buf_num_bytes_occupied; }
-
-private:
-    static constexpr size_t cMaxRecordSize = 16ULL * 1024 * 1024;
-
-    std::shared_ptr<ServerContext> m_server_ctx;
-    tcp::socket m_socket;
-    std::vector<char> m_buf;
-    size_t m_buf_num_bytes_occupied{0};
-};
-
 class AcceptTask {
 public:
     explicit AcceptTask(std::shared_ptr<RecordReceiverContext> rctx)
@@ -164,81 +85,6 @@ void queue_scheduler_update_listener_task(
 );
 void queue_validate_sender_task(std::shared_ptr<RecordReceiverContext> const& ctx);
 std::string server_status_to_string(ServerStatus status);
-
-bool RecordReceiverContext::read_connection_init_packet() {
-    job_id_t job_id{0};
-
-    if (m_buf_num_bytes_occupied != sizeof(job_id)) {
-        SPDLOG_ERROR("Rejecting connection due to invalid negotiation");
-        return false;
-    }
-
-    memcpy(&job_id, m_buf.data(), sizeof(job_id));
-    if (job_id != m_server_ctx->get_job_id()) {
-        SPDLOG_ERROR(
-                "Rejecting connection from worker with job_id={} during processing of "
-                "job_id={}",
-                job_id,
-                m_server_ctx->get_job_id()
-        );
-        return false;
-    }
-    m_buf_num_bytes_occupied = 0;
-
-    return true;
-}
-
-bool RecordReceiverContext::send_connection_accept_packet() {
-    char const response = 'y';
-    boost::system::error_code e;
-    auto transferred
-            = boost::asio::write(m_socket, boost::asio::buffer(&response, sizeof(response)), e);
-    if (e || transferred < sizeof(response)) {
-        SPDLOG_ERROR("Rejecting connection due to failure to send acceptance - {}", e.message());
-        return false;
-    }
-
-    return true;
-}
-
-bool RecordReceiverContext::read_record_groups_packet() {
-    size_t record_size{0};
-    auto* read_head = m_buf.data();
-    while (m_buf_num_bytes_occupied > 0) {
-        if (m_buf_num_bytes_occupied < sizeof(record_size)) {
-            break;
-        }
-        memcpy(&record_size, read_head, sizeof(record_size));
-
-        // terminate if record group size is over 16MB
-        if (record_size >= cMaxRecordSize) {
-            SPDLOG_ERROR("Record too large: {}B", record_size);
-            return false;
-        }
-
-        if (m_buf_num_bytes_occupied < record_size + sizeof(record_size)) {
-            break;
-        }
-        read_head += sizeof(record_size);
-
-        auto record_group = DeserializedRecordGroup{read_head, record_size};
-        m_server_ctx->push_record_group(record_group.get_tags(), record_group.record_iter());
-        m_buf_num_bytes_occupied -= (record_size + sizeof(record_size));
-        read_head += record_size;
-    }
-
-    if (m_buf_num_bytes_occupied > 0) {
-        if (m_buf.size() < record_size + sizeof(record_size)) {
-            std::vector<char> new_buf(sizeof(record_size) + record_size);
-            std::copy(m_buf.begin(), m_buf.end(), new_buf.begin());
-            m_buf.swap(new_buf);
-        } else {
-            memmove(m_buf.data(), read_head, m_buf_num_bytes_occupied);
-        }
-    }
-
-    return true;
-}
 
 void AcceptTask::operator()(boost::system::error_code const& error) {
     auto& server_ctx = m_record_recv_ctx->get_server_ctx();
